Fix code formatting and comments. Rename package remoteread to promread.
